functions/requirements.txt
azure-functions>=1.21.0
azure-functions-worker>=1.0.0
azure-identity>=1.17.1
azure-storage-blob>=12.22.0
azure-ai-formrecognizer>=3.3.3
anthropic>=0.40.0
python-pptx>=1.0.2
pyodbc>=5.1.0
fastapi>=0.115.0
pydantic>=2.9.0
PyJWT[crypto]>=2.9.0
httpx>=0.27.0
python-dateutil>=2.9.0
